requirements.txt
Flask==3.0.0
Requests==2.31.0
urllib3==2.1.0
python-dotenv==1.0.0
pymongo==4.6.1